+++ src/widgets/profile/profile.less.d.ts
@@ -1,0 +1,8 @@
+export interface IProfileLess {
+  'widget': string;
+  'row': string;
+  'title': string;
+  'chart': string;
+}
+
+export const locals: IProfileLess;